requirements.txt
python-dotenv
pybit
pandas
numpy
numba
pyyaml
